Moves source files into papertty package.
