+++ papertty/drivers/drivers_color.py
@@ -14,8 +14,9 @@
 #     You should have received a copy of the GNU General Public License
 #     along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-from drivers.drivers_full import WaveshareFull
+from papertty.drivers import WaveshareFull
 from PIL import Image
+
 
 
 class WaveshareColor(WaveshareFull):
