+++ papertty/drivers/drivers_full.py
@@ -15,7 +15,7 @@
 #     along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 from abc import abstractmethod
-from drivers.drivers_base import WaveshareEPD
+from papertty.drivers import WaveshareEPD
 
 
 class WaveshareFull(WaveshareEPD):
